Artillery changes
Making he MRA shooting salvos again
Buff Bombard and Pepperpot
Decreasing the Legged modifier for artillery

Tests showed that the MRA became a bit stronger than the Bombard. Therefore the Bombard is a bit buffed and to have still an upgrade when getting the Pepperpot, the Pepperpot too.
Decreased the Legged weaponModifier because from a logical point this modifier should not be higher than the modifer of the Anti-Personnel weapons.