Move Alpha 8 Base Transporter start location; Libcampaign fixes
Moves the Alpha 8 Base transporter to the second New Paradigm LZ location of Alpha 6.
This saves a lot of time as most players will have their good army fighting units up on the plateau for Alpha 7.

Allows setting transporter bases with a custom position rather than only computing the exact center of the cleanup area...

In doing so I fixed a few LZ coordinates which were not perfect tile multiples.